Resolve mixed option type
PHP 8 encourages using explicit types. OptionDefinition::$default can be null|bool|string currently. That's not possible to declare when using explicit types.

This change introduces the OptionValue interface with two implementations:
- BooleanOptionValue
- StringOptionValue

The idea to avoid mixed types is to return OptionValue or null, where OptionValue is a wrapper around a primitive type. The interface does not define any methods, because we only check the equality of these option values which can be done by the == operator on objects as well.