Do not expect all actions to provide help
Certain actions does not provide help option, but when checked by the ContextualHelpPlugin to decide whether to show the usage, it throws an exception currently:

InvalidArgumentException: The option "help" does not exist.

This change fixes that by thowing a domain specific exception NoSuchOptionException which is ignored by the ContextualHelpPlugin. Also it is now covered by a PHPUnit test case.